WIP: split kzg params getter
