+++ halo2-wasm/js/js/index.ts
@@ -2,7 +2,7 @@ import {
   Halo2Wasm, initPanicHook, Halo2LibWasm, CircuitConfig,
   Bn254FqPoint, Bn254G1AffinePoint, Bn254G2AffinePoint, JsCircuitBn254Fq2, JsCircuitBn254G1Affine, JsCircuitBn254G2Affine, JsCircuitSecp256k1Affine, JsCircuitValue256, Secp256k1AffinePoint
 } from "../../pkg/js/halo2_wasm";
-import { getKzgParams } from "../kzg";
+import { getKzgParams } from "./kzg";
 import { DEFAULT_CIRCUIT_CONFIG } from "../shared";
 import { BaseCircuitScaffold } from "../shared/scaffold";
