+++ halo2-wasm/js/shared/utils.ts
@@ -1,8 +1,4 @@
-import * as os from "os";
-import * as path from "path";
-import * as fs from "fs";
-
-function fetchAndConvertToUint8Array(url: string): Promise<Uint8Array> {
+export function fetchAndConvertToUint8Array(url: string): Promise<Uint8Array> {
   return new Promise((resolve, reject) => {
     // Check if running in Node.js environment
     if (
@@ -35,40 +31,11 @@ function fetchAndConvertToUint8Array(url: string): Promise<Uint8Array> {
   });
 }
 
-const convertBase64ToUint8Arr = (b64str: string) => {
+export const convertBase64ToUint8Arr = (b64str: string) => {
   const binstr = atob(b64str);
   const buf = new Uint8Array(binstr.length);
   Array.prototype.forEach.call(binstr, (ch, i) => {
     buf[i] = ch.charCodeAt(0);
   });
   return buf;
-};
-
-export const getKzgParams = async (k: number): Promise<Uint8Array> => {
-  const home = os.homedir();
-  const axiomSrsPath = path.join(
-    home,
-    ".axiom",
-    "srs",
-    "challenge_0085",
-    `kzg_bn254_${k}.srs`
-  );
-  const exists = fs.existsSync(axiomSrsPath);
-  if (exists) {
-    const buffer = fs.readFileSync(axiomSrsPath);
-    return new Uint8Array(buffer);
-  }
-  const folderPath = path.dirname(axiomSrsPath);
-  if (!fs.existsSync(folderPath)) {
-    fs.mkdirSync(folderPath, { recursive: true });
-  }
-  if (k < 6 || k > 19) {
-    throw new Error(`k=${k} is not supported`);
-  }
-  const srs = await fetchAndConvertToUint8Array(
-    `https://axiom-crypto.s3.amazonaws.com/challenge_0085/kzg_bn254_${k}.srs`
-  );
-  fs.writeFileSync(axiomSrsPath, srs);
-
-  return srs;
 };
